(MetallJsonLines) abstract functionality into class
Most of the functionality in MetallJsonLines was contained in the
C++ executables called form clippy/clippy-cpp.

This new version abstracts away most of the code into a separate
MetallJsonLines class. The C++ executables (e.g., mf-init, mf-count)
are responsible to interacting with Clippy, but after the arguments have
been read, the execution is deferred to the MetallJsonLines class.

The goal of the MetallJsonLines class is to model a distributed object,
whereas the goal of the C++ executables is to glue together the
clippy calls with the MetallJsonLines object.

Note: mf-merge has not been fully converted.