(MetallGraph) Added first version of MetallGraph
A MetallGraph contains two MetallJsonLines objects,
that share a Metall MPI adaptor / memory manager.

Refactored the Metall MPI adaptor / memory manager out of the
MetallJsonLines class, and made it a responsibility of the
user to create it first.
This way, the memory manager can be shared across multiple
data structures within the same Metall store.